refactor(): change project to esm module
Project changed to use the new ESM module engine
Upgraded all dependencies
