+++ package.json
@@ -2,9 +2,10 @@
   "name": "ls-node-cli-template",
   "version": "1.0.0",
   "description": "Add a description here",
-  "main": "index.js",
+  "exports": "./index.js",
+  "type": "module",
   "scripts": {
-    "start": "ts-node --project tsconfig.json index.ts",
+    "start": "ts-node-esm --project tsconfig.json index.ts",
     "lint": "./node_modules/typescript/bin/tsc --project tsconfig.json --noEmit && eslint . --ext .js,.ts",
     "lint:fix": "./node_modules/tpescript/tsc --project tsconfig.json --noEmit && eslint . --ext .js,.ts --fix",
     "prepare": "husky install"
@@ -12,35 +13,43 @@
   "keywords": [],
   "author": "",
   "license": "ISC",
+  "engines": {
+    "node": ">=16.0.0",
+    "npm": ">=8.0.0"
+  },
   "dependencies": {
-    "chalk": "^4.1.2",
-    "inquirer": "^8.2.5",
-    "ora": "^5.4.1",
-    "yargs": "^17.5.1"
+    "@inquirer/confirm": "^1.0.11",
+    "@inquirer/prompts": "^1.2.3",
+    "chalk": "^5.2.0",
+    "ora": "^6.3.1",
+    "yargs": "^17.7.2"
   },
   "devDependencies": {
-    "@commitlint/cli": "^17.3.0",
-    "@commitlint/config-conventional": "^17.3.0",
-    "@types/inquirer": "^8.2.5",
-    "@types/node": "^18.0.4",
+    "@commitlint/cli": "^17.6.3",
+    "@commitlint/config-conventional": "^17.6.3",
+    "@types/glob": "^8.1.0",
+    "@types/inquirer": "^9.0.3",
+    "@types/node": "^20.2.4",
+    "@types/node-fetch": "^2.6.4",
     "@types/ora": "^3.2.0",
-    "@types/yargs": "^17.0.10",
-    "@typescript-eslint/eslint-plugin": "^5.30.6",
-    "@typescript-eslint/parser": "^5.30.6",
-    "eslint": "^8.19.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-import-resolver-node": "^0.3.6",
-    "eslint-import-resolver-typescript": "^3.2.6",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-import-helpers": "^1.2.1",
+    "@types/p-queue": "^3.2.1",
+    "@types/yargs": "^17.0.24",
+    "@typescript-eslint/eslint-plugin": "^5.59.7",
+    "@typescript-eslint/parser": "^5.59.7",
+    "eslint": "^8.41.0",
+    "eslint-config-prettier": "^8.8.0",
+    "eslint-import-resolver-node": "^0.3.7",
+    "eslint-import-resolver-typescript": "^3.5.5",
+    "eslint-plugin-import": "^2.27.5",
+    "eslint-plugin-import-helpers": "^1.3.1",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-promise": "^6.0.0",
+    "eslint-plugin-promise": "^6.1.1",
     "eslint-plugin-unused-imports": "2.0.0",
-    "husky": "^8.0.2",
-    "prettier": "^2.7.1",
+    "husky": "^8.0.3",
+    "prettier": "^2.8.8",
     "ts-node": "^10.9.1",
-    "tsconfig-paths": "^4.0.0",
-    "typescript": "^4.7.4"
+    "tsconfig-paths": "^4.2.0",
+    "typescript": "^5.0.4"
   }
 }
