+++ src/services/index.ts
@@ -1,2 +1,2 @@
-export { default as args } from './args'
-export { default as questions } from './questions'
+export { default as args } from './args/index.js'
+export { default as questions } from './questions/index.js'
