+++ tsconfig.json
@@ -1,7 +1,7 @@
 {
     "compilerOptions": {
-        "module": "commonjs",
-        "moduleResolution": "node",
+        "module": "Node16",
+        "moduleResolution": "node16",
         "pretty": true,
         "sourceMap": true,
         "target": "ESNext",
@@ -25,5 +25,10 @@
     ],
     "exclude": [
         "node_modules"
-    ]
+    ],
+    "ts-node": {
+        "transpileOnly": true,
+        "files": true,
+        "experimentalResolver": true,
+    }
 }
